@@ -6,7 +6,8 @@
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
-    "lint": "npx ng lint --fix && npm run lint:styles",
+    "lint": "npm run lint:ts && npm run lint:styles",
+    "lint:ts": "npx ng lint --fix",
     "lint:styles": "stylelint --fix src/**/*.scss",
     "e2e": "ng e2e",
     "audit:ci": "npm audit --json|npm-audit-reporter-teamcity",
@@ -29,7 +30,7 @@
     "@angular/platform-browser": "^12.1.3",
     "@angular/platform-browser-dynamic": "^12.1.3",
     "@angular/router": "^12.1.3",
-    "@enigmatry/angular-building-blocks": "^1.0.20",
+    "@enigmatry/angular-building-blocks": "^1.0.44",
     "@ngx-formly/core": "^5.10.22",
     "@ngx-formly/material": "^5.10.22",
     "nswag": "^13.12.1",
@@ -39,28 +40,13 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^12.1.3",
-    "@angular-eslint/builder": "^12.3.1",
-    "@angular-eslint/eslint-plugin": "^12.3.1",
-    "@angular-eslint/eslint-plugin-template": "^12.3.1",
-    "@angular-eslint/schematics": "^12.3.1",
-    "@angular-eslint/template-parser": "^12.3.1",
     "@angular/cli": "^12.1.3",
     "@angular/compiler-cli": "^12.1.3",
     "@angular/localize": "^12.1.3",
     "@types/jasmine": "^3.8.1",
     "@types/node": "^16.4.3",
-    "@typescript-eslint/eslint-plugin": "4.28.4",
-    "@typescript-eslint/parser": "4.28.4",
-    "eslint": "^7.31.0",
-    "eslint-plugin-import": "2.23.4",
-    "eslint-plugin-jsdoc": "35.5.1",
-    "eslint-plugin-no-loops": "0.3.0",
-    "eslint-plugin-no-secrets": "0.8.9",
-    "eslint-plugin-no-unsanitized": "3.1.5",
-    "eslint-plugin-prefer-arrow": "1.2.3",
-    "eslint-plugin-promise": "5.1.0",
-    "eslint-plugin-unused-imports": "1.1.1",
-    "eslint-plugin-xss": "0.1.10",
+    "eslint": "^7.32.0",
+    "@enigmatry/eslint-config": "^1.0.0",
     "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.3.4",
@@ -71,13 +57,7 @@
     "npm-audit-reporter-teamcity": "0.3.1",
     "protractor": "~7.0.0",
     "stylelint": "13.13.1",
-    "stylelint-config-property-sort-order-smacss": "7.1.0",
-    "stylelint-declaration-block-no-ignored-properties": "2.4.0",
-    "stylelint-group-selectors": "1.0.8",
-    "stylelint-max-lines": "0.1.0",
-    "stylelint-scss": "3.20.1",
-    "stylelint-selector-no-empty": "1.0.8",
-    "stylelint-use-nesting": "3.0.0",
+    "@enigmatry/stylelint-config": "^1.0.0",
     "ts-node": "~10.1.0",
     "typescript": "4.3.5"
   }
