@@ -17,7 +17,7 @@ namespace Enigmatry.Blueprint.CodeGeneration.Setup.Features.Users
             builder.FormControl(x => x.CreatedOn).IsReadonly(true);
             builder.FormControl(x => x.UpdatedOn).IsReadonly(true);
 
-            builder.WithValidationConfiguration(new UserEditValidationConfiguration());
+            builder.WithValidationConfiguration(new UserEditComponentValidationConfiguration());
         }
     }
 }
