@@ -17,7 +17,7 @@ namespace Enigmatry.Blueprint.Data.Migrations.Seeding
             });
 
             user.SetCreated(new DateTimeOffset(2019, 5, 6, 14, 31, 0, TimeSpan.FromHours(0)));
-            modelBuilder.Entity<User>().HasData(user.WithId(new Guid("8207DB25-94D1-4F3D-BF18-90DA283221F7")));
+            modelBuilder.Entity<User>().HasData(user.WithId(User.TestUserId));
         }
     }
 }
