@@ -16,9 +16,10 @@ namespace Enigmatry.Blueprint.Infrastructure.Data.Configurations
 
             builder.HasIndex(x => x.UserName).IsUnique();
 
-
             builder.HasMany(x => x.CreatedUsers).WithOne(x => x.CreatedBy!);
             builder.HasMany(x => x.UpdatedUsers).WithOne(x => x.UpdatedBy!);
+            builder.HasMany(x => x.CreatedProducts).WithOne(x => x.CreatedBy!);
+            builder.HasMany(x => x.UpdatedProducts).WithOne(x => x.UpdatedBy!);
         }
     }
 }
