@@ -2,7 +2,7 @@
 /* eslint-disable */
 //----------------------
 // <auto-generated>
-//     Generated using the NSwag toolchain v13.11.3.0 (NJsonSchema v10.4.4.0 (Newtonsoft.Json v11.0.0.0)) (http://NSwag.org)
+//     Generated using the NSwag toolchain v13.13.2.0 (NJsonSchema v10.5.2.0 (Newtonsoft.Json v11.0.0.0)) (http://NSwag.org)
 // </auto-generated>
 //----------------------
 // ReSharper disable InconsistentNaming
@@ -309,6 +309,380 @@ export class UsersClient implements IUsersClient {
     }
 }
 
+export interface IProductsClient {
+    /**
+     * Gets listing of all available users
+     * @param keyword (optional) 
+     * @param pageNumber (optional) 
+     * @param pageSize (optional) 
+     * @param sortBy (optional) 
+     * @param sortDirection (optional) 
+     * @return List of products
+     */
+    search(keyword: string | null | undefined, pageNumber: number | undefined, pageSize: number | undefined, sortBy: string | null | undefined, sortDirection: string | null | undefined): Observable<PagedResponseOfGetProductsResponseItem>;
+    /**
+     * Creates or updates
+     * @param command Product data
+     */
+    post(command: ProductCreateOrUpdateCommand): Observable<ProductCreateOrUpdateResult>;
+    /**
+     * Get product for given id
+     * @param id Id
+     */
+    get(id: string): Observable<GetProductDetailsResponse>;
+    /**
+     * Return true if product code is unique
+     * @param code Code
+     */
+    isCodeUnique(code: string | null): Observable<GetProductCodeUniquenessResponse>;
+    /**
+     * Return true if product name is unique
+     * @param name Name
+     */
+    isNameUnique(name: string | null): Observable<GetProductNameUniquenessResponse>;
+}
+
+@Injectable({
+    providedIn: 'root'
+})
+export class ProductsClient implements IProductsClient {
+    private http: HttpClient;
+    private baseUrl: string;
+    protected jsonParseReviver: ((key: string, value: any) => any) | undefined = undefined;
+
+    constructor(@Inject(HttpClient) http: HttpClient, @Optional() @Inject(API_BASE_URL) baseUrl?: string) {
+        this.http = http;
+        this.baseUrl = baseUrl !== undefined && baseUrl !== null ? baseUrl : "http://localhost:5440";
+    }
+
+    /**
+     * Gets listing of all available users
+     * @param keyword (optional) 
+     * @param pageNumber (optional) 
+     * @param pageSize (optional) 
+     * @param sortBy (optional) 
+     * @param sortDirection (optional) 
+     * @return List of products
+     */
+    search(keyword: string | null | undefined, pageNumber: number | undefined, pageSize: number | undefined, sortBy: string | null | undefined, sortDirection: string | null | undefined): Observable<PagedResponseOfGetProductsResponseItem> {
+        let url_ = this.baseUrl + "/Products?";
+        if (keyword !== undefined && keyword !== null)
+            url_ += "Keyword=" + encodeURIComponent("" + keyword) + "&";
+        if (pageNumber === null)
+            throw new Error("The parameter 'pageNumber' cannot be null.");
+        else if (pageNumber !== undefined)
+            url_ += "PageNumber=" + encodeURIComponent("" + pageNumber) + "&";
+        if (pageSize === null)
+            throw new Error("The parameter 'pageSize' cannot be null.");
+        else if (pageSize !== undefined)
+            url_ += "PageSize=" + encodeURIComponent("" + pageSize) + "&";
+        if (sortBy !== undefined && sortBy !== null)
+            url_ += "SortBy=" + encodeURIComponent("" + sortBy) + "&";
+        if (sortDirection !== undefined && sortDirection !== null)
+            url_ += "SortDirection=" + encodeURIComponent("" + sortDirection) + "&";
+        url_ = url_.replace(/[?&]$/, "");
+
+        let options_ : any = {
+            observe: "response",
+            responseType: "blob",
+            headers: new HttpHeaders({
+                "Accept": "application/json"
+            })
+        };
+
+        return this.http.request("get", url_, options_).pipe(_observableMergeMap((response_ : any) => {
+            return this.processSearch(response_);
+        })).pipe(_observableCatch((response_: any) => {
+            if (response_ instanceof HttpResponseBase) {
+                try {
+                    return this.processSearch(<any>response_);
+                } catch (e) {
+                    return <Observable<PagedResponseOfGetProductsResponseItem>><any>_observableThrow(e);
+                }
+            } else
+                return <Observable<PagedResponseOfGetProductsResponseItem>><any>_observableThrow(response_);
+        }));
+    }
+
+    protected processSearch(response: HttpResponseBase): Observable<PagedResponseOfGetProductsResponseItem> {
+        const status = response.status;
+        const responseBlob =
+            response instanceof HttpResponse ? response.body :
+            (<any>response).error instanceof Blob ? (<any>response).error : undefined;
+
+        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }}
+        if (status === 200) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result200: any = null;
+            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result200 = PagedResponseOfGetProductsResponseItem.fromJS(resultData200);
+            return _observableOf(result200);
+            }));
+        } else if (status !== 200 && status !== 204) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
+            }));
+        }
+        return _observableOf<PagedResponseOfGetProductsResponseItem>(<any>null);
+    }
+
+    /**
+     * Creates or updates
+     * @param command Product data
+     */
+    post(command: ProductCreateOrUpdateCommand): Observable<ProductCreateOrUpdateResult> {
+        let url_ = this.baseUrl + "/Products";
+        url_ = url_.replace(/[?&]$/, "");
+
+        const content_ = JSON.stringify(command);
+
+        let options_ : any = {
+            body: content_,
+            observe: "response",
+            responseType: "blob",
+            headers: new HttpHeaders({
+                "Content-Type": "application/json",
+                "Accept": "application/json"
+            })
+        };
+
+        return this.http.request("post", url_, options_).pipe(_observableMergeMap((response_ : any) => {
+            return this.processPost(response_);
+        })).pipe(_observableCatch((response_: any) => {
+            if (response_ instanceof HttpResponseBase) {
+                try {
+                    return this.processPost(<any>response_);
+                } catch (e) {
+                    return <Observable<ProductCreateOrUpdateResult>><any>_observableThrow(e);
+                }
+            } else
+                return <Observable<ProductCreateOrUpdateResult>><any>_observableThrow(response_);
+        }));
+    }
+
+    protected processPost(response: HttpResponseBase): Observable<ProductCreateOrUpdateResult> {
+        const status = response.status;
+        const responseBlob =
+            response instanceof HttpResponse ? response.body :
+            (<any>response).error instanceof Blob ? (<any>response).error : undefined;
+
+        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }}
+        if (status === 200) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result200: any = null;
+            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result200 = ProductCreateOrUpdateResult.fromJS(resultData200);
+            return _observableOf(result200);
+            }));
+        } else if (status === 404) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result404: any = null;
+            let resultData404 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result404 = ProblemDetails.fromJS(resultData404);
+            return throwException("A server side error occurred.", status, _responseText, _headers, result404);
+            }));
+        } else if (status === 400) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result400: any = null;
+            let resultData400 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result400 = ProblemDetails.fromJS(resultData400);
+            return throwException("A server side error occurred.", status, _responseText, _headers, result400);
+            }));
+        } else if (status !== 200 && status !== 204) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
+            }));
+        }
+        return _observableOf<ProductCreateOrUpdateResult>(<any>null);
+    }
+
+    /**
+     * Get product for given id
+     * @param id Id
+     */
+    get(id: string): Observable<GetProductDetailsResponse> {
+        let url_ = this.baseUrl + "/Products/{id}";
+        if (id === undefined || id === null)
+            throw new Error("The parameter 'id' must be defined.");
+        url_ = url_.replace("{id}", encodeURIComponent("" + id));
+        url_ = url_.replace(/[?&]$/, "");
+
+        let options_ : any = {
+            observe: "response",
+            responseType: "blob",
+            headers: new HttpHeaders({
+                "Accept": "application/json"
+            })
+        };
+
+        return this.http.request("get", url_, options_).pipe(_observableMergeMap((response_ : any) => {
+            return this.processGet(response_);
+        })).pipe(_observableCatch((response_: any) => {
+            if (response_ instanceof HttpResponseBase) {
+                try {
+                    return this.processGet(<any>response_);
+                } catch (e) {
+                    return <Observable<GetProductDetailsResponse>><any>_observableThrow(e);
+                }
+            } else
+                return <Observable<GetProductDetailsResponse>><any>_observableThrow(response_);
+        }));
+    }
+
+    protected processGet(response: HttpResponseBase): Observable<GetProductDetailsResponse> {
+        const status = response.status;
+        const responseBlob =
+            response instanceof HttpResponse ? response.body :
+            (<any>response).error instanceof Blob ? (<any>response).error : undefined;
+
+        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }}
+        if (status === 200) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result200: any = null;
+            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result200 = GetProductDetailsResponse.fromJS(resultData200);
+            return _observableOf(result200);
+            }));
+        } else if (status === 404) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result404: any = null;
+            let resultData404 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result404 = ProblemDetails.fromJS(resultData404);
+            return throwException("A server side error occurred.", status, _responseText, _headers, result404);
+            }));
+        } else if (status !== 200 && status !== 204) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
+            }));
+        }
+        return _observableOf<GetProductDetailsResponse>(<any>null);
+    }
+
+    /**
+     * Return true if product code is unique
+     * @param code Code
+     */
+    isCodeUnique(code: string | null): Observable<GetProductCodeUniquenessResponse> {
+        let url_ = this.baseUrl + "/Products/code-unique/{code}";
+        if (code === undefined || code === null)
+            throw new Error("The parameter 'code' must be defined.");
+        url_ = url_.replace("{code}", encodeURIComponent("" + code));
+        url_ = url_.replace(/[?&]$/, "");
+
+        let options_ : any = {
+            observe: "response",
+            responseType: "blob",
+            headers: new HttpHeaders({
+                "Accept": "application/json"
+            })
+        };
+
+        return this.http.request("get", url_, options_).pipe(_observableMergeMap((response_ : any) => {
+            return this.processIsCodeUnique(response_);
+        })).pipe(_observableCatch((response_: any) => {
+            if (response_ instanceof HttpResponseBase) {
+                try {
+                    return this.processIsCodeUnique(<any>response_);
+                } catch (e) {
+                    return <Observable<GetProductCodeUniquenessResponse>><any>_observableThrow(e);
+                }
+            } else
+                return <Observable<GetProductCodeUniquenessResponse>><any>_observableThrow(response_);
+        }));
+    }
+
+    protected processIsCodeUnique(response: HttpResponseBase): Observable<GetProductCodeUniquenessResponse> {
+        const status = response.status;
+        const responseBlob =
+            response instanceof HttpResponse ? response.body :
+            (<any>response).error instanceof Blob ? (<any>response).error : undefined;
+
+        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }}
+        if (status === 200) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result200: any = null;
+            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result200 = GetProductCodeUniquenessResponse.fromJS(resultData200);
+            return _observableOf(result200);
+            }));
+        } else if (status === 404) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result404: any = null;
+            let resultData404 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result404 = ProblemDetails.fromJS(resultData404);
+            return throwException("A server side error occurred.", status, _responseText, _headers, result404);
+            }));
+        } else if (status !== 200 && status !== 204) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
+            }));
+        }
+        return _observableOf<GetProductCodeUniquenessResponse>(<any>null);
+    }
+
+    /**
+     * Return true if product name is unique
+     * @param name Name
+     */
+    isNameUnique(name: string | null): Observable<GetProductNameUniquenessResponse> {
+        let url_ = this.baseUrl + "/Products/name-unique/{name}";
+        if (name === undefined || name === null)
+            throw new Error("The parameter 'name' must be defined.");
+        url_ = url_.replace("{name}", encodeURIComponent("" + name));
+        url_ = url_.replace(/[?&]$/, "");
+
+        let options_ : any = {
+            observe: "response",
+            responseType: "blob",
+            headers: new HttpHeaders({
+                "Accept": "application/json"
+            })
+        };
+
+        return this.http.request("get", url_, options_).pipe(_observableMergeMap((response_ : any) => {
+            return this.processIsNameUnique(response_);
+        })).pipe(_observableCatch((response_: any) => {
+            if (response_ instanceof HttpResponseBase) {
+                try {
+                    return this.processIsNameUnique(<any>response_);
+                } catch (e) {
+                    return <Observable<GetProductNameUniquenessResponse>><any>_observableThrow(e);
+                }
+            } else
+                return <Observable<GetProductNameUniquenessResponse>><any>_observableThrow(response_);
+        }));
+    }
+
+    protected processIsNameUnique(response: HttpResponseBase): Observable<GetProductNameUniquenessResponse> {
+        const status = response.status;
+        const responseBlob =
+            response instanceof HttpResponse ? response.body :
+            (<any>response).error instanceof Blob ? (<any>response).error : undefined;
+
+        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }}
+        if (status === 200) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result200: any = null;
+            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result200 = GetProductNameUniquenessResponse.fromJS(resultData200);
+            return _observableOf(result200);
+            }));
+        } else if (status === 404) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result404: any = null;
+            let resultData404 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result404 = ProblemDetails.fromJS(resultData404);
+            return throwException("A server side error occurred.", status, _responseText, _headers, result404);
+            }));
+        } else if (status !== 200 && status !== 204) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
+            }));
+        }
+        return _observableOf<GetProductNameUniquenessResponse>(<any>null);
+    }
+}
+
 export class PagedResponseOfGetUsersResponseItem implements IPagedResponseOfGetUsersResponseItem {
     items?: GetUsersResponseItem[];
     totalCount?: number;
@@ -481,12 +855,27 @@ export interface IGetUserDetailsResponse {
     updatedOn?: Date;
 }
 
+/** A machine-readable format for specifying errors in HTTP API responses based on https://tools.ietf.org/html/rfc7807. */
 export class ProblemDetails implements IProblemDetails {
+    /** A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when
+dereferenced, it provide human-readable documentation for the problem type
+(e.g., using HTML [W3C.REC-html5-20141028]).  When this member is not present, its value is assumed to be
+"about:blank". */
     type?: string | undefined;
+    /** A short, human-readable summary of the problem type.It SHOULD NOT change from occurrence to occurrence
+of the problem, except for purposes of localization(e.g., using proactive content negotiation;
+see[RFC7231], Section 3.4). */
     title?: string | undefined;
+    /** The HTTP status code([RFC7231], Section 6) generated by the origin server for this occurrence of the problem. */
     status?: number | undefined;
+    /** A human-readable explanation specific to this occurrence of the problem. */
     detail?: string | undefined;
+    /** A URI reference that identifies the specific occurrence of the problem.It may or may not yield further information if dereferenced. */
     instance?: string | undefined;
+    /** Gets the IDictionary`2 for extension members.
+
+Problem type definitions MAY extend the problem details object with additional members. Extension members appear in the same namespace as
+other members of a problem type. */
     extensions?: { [key: string]: any; } | undefined;
 
     constructor(data?: IProblemDetails) {
@@ -540,12 +929,27 @@ export class ProblemDetails implements IProblemDetails {
     }
 }
 
+/** A machine-readable format for specifying errors in HTTP API responses based on https://tools.ietf.org/html/rfc7807. */
 export interface IProblemDetails {
+    /** A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when
+dereferenced, it provide human-readable documentation for the problem type
+(e.g., using HTML [W3C.REC-html5-20141028]).  When this member is not present, its value is assumed to be
+"about:blank". */
     type?: string | undefined;
+    /** A short, human-readable summary of the problem type.It SHOULD NOT change from occurrence to occurrence
+of the problem, except for purposes of localization(e.g., using proactive content negotiation;
+see[RFC7231], Section 3.4). */
     title?: string | undefined;
+    /** The HTTP status code([RFC7231], Section 6) generated by the origin server for this occurrence of the problem. */
     status?: number | undefined;
+    /** A human-readable explanation specific to this occurrence of the problem. */
     detail?: string | undefined;
+    /** A URI reference that identifies the specific occurrence of the problem.It may or may not yield further information if dereferenced. */
     instance?: string | undefined;
+    /** Gets the IDictionary`2 for extension members.
+
+Problem type definitions MAY extend the problem details object with additional members. Extension members appear in the same namespace as
+other members of a problem type. */
     extensions?: { [key: string]: any; } | undefined;
 }
 
@@ -591,6 +995,405 @@ export interface IUserCreateOrUpdateCommand {
     id?: string | undefined;
     userName?: string;
     name?: string;
+}
+
+export class PagedResponseOfGetProductsResponseItem implements IPagedResponseOfGetProductsResponseItem {
+    items?: GetProductsResponseItem[];
+    totalCount?: number;
+    pageNumber?: number;
+    pageSize?: number;
+    totalPages?: number;
+    hasNextPage?: boolean;
+    hasPreviousPage?: boolean;
+
+    constructor(data?: IPagedResponseOfGetProductsResponseItem) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            if (Array.isArray(_data["items"])) {
+                this.items = [] as any;
+                for (let item of _data["items"])
+                    this.items!.push(GetProductsResponseItem.fromJS(item));
+            }
+            this.totalCount = _data["totalCount"];
+            this.pageNumber = _data["pageNumber"];
+            this.pageSize = _data["pageSize"];
+            this.totalPages = _data["totalPages"];
+            this.hasNextPage = _data["hasNextPage"];
+            this.hasPreviousPage = _data["hasPreviousPage"];
+        }
+    }
+
+    static fromJS(data: any): PagedResponseOfGetProductsResponseItem {
+        data = typeof data === 'object' ? data : {};
+        let result = new PagedResponseOfGetProductsResponseItem();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        if (Array.isArray(this.items)) {
+            data["items"] = [];
+            for (let item of this.items)
+                data["items"].push(item.toJSON());
+        }
+        data["totalCount"] = this.totalCount;
+        data["pageNumber"] = this.pageNumber;
+        data["pageSize"] = this.pageSize;
+        data["totalPages"] = this.totalPages;
+        data["hasNextPage"] = this.hasNextPage;
+        data["hasPreviousPage"] = this.hasPreviousPage;
+        return data; 
+    }
+}
+
+export interface IPagedResponseOfGetProductsResponseItem {
+    items?: GetProductsResponseItem[];
+    totalCount?: number;
+    pageNumber?: number;
+    pageSize?: number;
+    totalPages?: number;
+    hasNextPage?: boolean;
+    hasPreviousPage?: boolean;
+}
+
+export class GetProductsResponseItem implements IGetProductsResponseItem {
+    id?: string;
+    name?: string;
+    code?: string;
+    type?: ProductType;
+    price?: number;
+    amount?: number;
+    contactEmail?: string;
+    contactPhone?: string;
+    infoLink?: string;
+    expiresOn?: Date | undefined;
+
+    constructor(data?: IGetProductsResponseItem) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+            this.name = _data["name"];
+            this.code = _data["code"];
+            this.type = _data["type"];
+            this.price = _data["price"];
+            this.amount = _data["amount"];
+            this.contactEmail = _data["contactEmail"];
+            this.contactPhone = _data["contactPhone"];
+            this.infoLink = _data["infoLink"];
+            this.expiresOn = _data["expiresOn"] ? new Date(_data["expiresOn"].toString()) : <any>undefined;
+        }
+    }
+
+    static fromJS(data: any): GetProductsResponseItem {
+        data = typeof data === 'object' ? data : {};
+        let result = new GetProductsResponseItem();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["id"] = this.id;
+        data["name"] = this.name;
+        data["code"] = this.code;
+        data["type"] = this.type;
+        data["price"] = this.price;
+        data["amount"] = this.amount;
+        data["contactEmail"] = this.contactEmail;
+        data["contactPhone"] = this.contactPhone;
+        data["infoLink"] = this.infoLink;
+        data["expiresOn"] = this.expiresOn ? this.expiresOn.toISOString() : <any>undefined;
+        return data; 
+    }
+}
+
+export interface IGetProductsResponseItem {
+    id?: string;
+    name?: string;
+    code?: string;
+    type?: ProductType;
+    price?: number;
+    amount?: number;
+    contactEmail?: string;
+    contactPhone?: string;
+    infoLink?: string;
+    expiresOn?: Date | undefined;
+}
+
+export enum ProductType {
+    Food = 0,
+    Drink = 1,
+    Book = 2,
+    Car = 3,
+}
+
+export class GetProductDetailsResponse implements IGetProductDetailsResponse {
+    id?: string;
+    name?: string;
+    code?: string;
+    type?: ProductType;
+    price?: number;
+    amount?: number;
+    contactEmail?: string;
+    contactPhone?: string;
+    infoLink?: string;
+    expiresOn?: Date | undefined;
+
+    constructor(data?: IGetProductDetailsResponse) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+            this.name = _data["name"];
+            this.code = _data["code"];
+            this.type = _data["type"];
+            this.price = _data["price"];
+            this.amount = _data["amount"];
+            this.contactEmail = _data["contactEmail"];
+            this.contactPhone = _data["contactPhone"];
+            this.infoLink = _data["infoLink"];
+            this.expiresOn = _data["expiresOn"] ? new Date(_data["expiresOn"].toString()) : <any>undefined;
+        }
+    }
+
+    static fromJS(data: any): GetProductDetailsResponse {
+        data = typeof data === 'object' ? data : {};
+        let result = new GetProductDetailsResponse();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["id"] = this.id;
+        data["name"] = this.name;
+        data["code"] = this.code;
+        data["type"] = this.type;
+        data["price"] = this.price;
+        data["amount"] = this.amount;
+        data["contactEmail"] = this.contactEmail;
+        data["contactPhone"] = this.contactPhone;
+        data["infoLink"] = this.infoLink;
+        data["expiresOn"] = this.expiresOn ? this.expiresOn.toISOString() : <any>undefined;
+        return data; 
+    }
+}
+
+export interface IGetProductDetailsResponse {
+    id?: string;
+    name?: string;
+    code?: string;
+    type?: ProductType;
+    price?: number;
+    amount?: number;
+    contactEmail?: string;
+    contactPhone?: string;
+    infoLink?: string;
+    expiresOn?: Date | undefined;
+}
+
+export class GetProductCodeUniquenessResponse implements IGetProductCodeUniquenessResponse {
+    isUnique?: boolean;
+
+    constructor(data?: IGetProductCodeUniquenessResponse) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.isUnique = _data["isUnique"];
+        }
+    }
+
+    static fromJS(data: any): GetProductCodeUniquenessResponse {
+        data = typeof data === 'object' ? data : {};
+        let result = new GetProductCodeUniquenessResponse();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["isUnique"] = this.isUnique;
+        return data; 
+    }
+}
+
+export interface IGetProductCodeUniquenessResponse {
+    isUnique?: boolean;
+}
+
+export class GetProductNameUniquenessResponse implements IGetProductNameUniquenessResponse {
+    isUnique?: boolean;
+
+    constructor(data?: IGetProductNameUniquenessResponse) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.isUnique = _data["isUnique"];
+        }
+    }
+
+    static fromJS(data: any): GetProductNameUniquenessResponse {
+        data = typeof data === 'object' ? data : {};
+        let result = new GetProductNameUniquenessResponse();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["isUnique"] = this.isUnique;
+        return data; 
+    }
+}
+
+export interface IGetProductNameUniquenessResponse {
+    isUnique?: boolean;
+}
+
+export class ProductCreateOrUpdateResult implements IProductCreateOrUpdateResult {
+    id?: string;
+
+    constructor(data?: IProductCreateOrUpdateResult) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+        }
+    }
+
+    static fromJS(data: any): ProductCreateOrUpdateResult {
+        data = typeof data === 'object' ? data : {};
+        let result = new ProductCreateOrUpdateResult();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["id"] = this.id;
+        return data; 
+    }
+}
+
+export interface IProductCreateOrUpdateResult {
+    id?: string;
+}
+
+export class ProductCreateOrUpdateCommand implements IProductCreateOrUpdateCommand {
+    id?: string | undefined;
+    name?: string;
+    code?: string;
+    type?: ProductType;
+    price?: number;
+    amount?: number;
+    contactEmail?: string;
+    contactPhone?: string;
+    infoLink?: string;
+    expiresOn?: Date | undefined;
+
+    constructor(data?: IProductCreateOrUpdateCommand) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+            this.name = _data["name"];
+            this.code = _data["code"];
+            this.type = _data["type"];
+            this.price = _data["price"];
+            this.amount = _data["amount"];
+            this.contactEmail = _data["contactEmail"];
+            this.contactPhone = _data["contactPhone"];
+            this.infoLink = _data["infoLink"];
+            this.expiresOn = _data["expiresOn"] ? new Date(_data["expiresOn"].toString()) : <any>undefined;
+        }
+    }
+
+    static fromJS(data: any): ProductCreateOrUpdateCommand {
+        data = typeof data === 'object' ? data : {};
+        let result = new ProductCreateOrUpdateCommand();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["id"] = this.id;
+        data["name"] = this.name;
+        data["code"] = this.code;
+        data["type"] = this.type;
+        data["price"] = this.price;
+        data["amount"] = this.amount;
+        data["contactEmail"] = this.contactEmail;
+        data["contactPhone"] = this.contactPhone;
+        data["infoLink"] = this.infoLink;
+        data["expiresOn"] = this.expiresOn ? this.expiresOn.toISOString() : <any>undefined;
+        return data; 
+    }
+}
+
+export interface IProductCreateOrUpdateCommand {
+    id?: string | undefined;
+    name?: string;
+    code?: string;
+    type?: ProductType;
+    price?: number;
+    amount?: number;
+    contactEmail?: string;
+    contactPhone?: string;
+    infoLink?: string;
+    expiresOn?: Date | undefined;
 }
 
 export class ApiException extends Error {
